// frontend/src/api/generated/models/CategoryRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Просмотр категории.
 */
export type CategoryRead = {
    id: string;
    name: string;
    slug: string;
    description: (string | null);
    is_active: boolean;
    created_at: string;
    updated_at: string;
};